Time: 6311 ms (22.45%), Space: 70.2 MB (10.2%) - LeetHub

+++ 3510-minimum-pair-removal-to-sort-array-ii/3510-minimum-pair-removal-to-sort-array-ii.py
@@ -6,7 +6,7 @@ class Solution(object):
         if n < 2:
             return 0
             
-        arr = [long(x) for x in nums]
+        arr = [int(x) for x in nums]
         left_idx = [i - 1 for i in range(n)]
         right_idx = [i + 1 for i in range(n)]
